requirements.txt
Flask
Werkzeug
vanna
duckdb
pandas
numpy
python-dotenv